backend/requirements.txt
fastapi
uvicorn
google-genai>=0.3.0
python-dotenv
pathspec
requests
